Handles "bell" and "page", image update hack
* The image has been updated with a bug fix, but it is more of a hack,
some value is being overwritten during image creation. Commit hash of
fix is: 155fc45f2725461378ab480cdc9383872872e875
* The "bell" and "page" forth words work, just enough of the ANSI escape
sequences were implemented to implement "page" (it is done in a super
inefficient way, a state-machine to parse out the escape sequences would
be better, as I have done before, but this "solution" is much simpler
and the project is just a toy). "bell" has been implemented as a series
of setTimeout() calls, it seems to work well.